Refactor sandbox to separate module
