--- .github/workflows/test-publish.yml
@@ -1,44 +1,0 @@
-name: Test Publish Workflow
-
-on:
-  workflow_dispatch:  # Manual trigger
-
-jobs:
-  build:
-    name: Test Build distribution 📦
-    runs-on: ubuntu-latest
-
-    steps:
-        - uses: actions/checkout@v4
-        - name: Set up Python
-          uses: actions/setup-python@v5
-          with:
-            python-version: "3.11"  # Use a version compatible with >=3.8,<3.12
-        - name: Install Poetry
-          uses: snok/install-poetry@v1
-          with:
-            version: 1.8.2  # Match local version
-        - name: Bump version
-          run: poetry version $(git describe --tags --abbrev=0)
-        - name: Build a binary wheel and a source tarball
-          run: poetry build
-        - name: Store the distribution packages
-          uses: actions/upload-artifact@v4
-          with:
-            name: python-package-distributions-test
-            path: dist/
-
-  # This step simulates the PyPI publish step without actually publishing
-  simulate-publish:
-    name: Simulate PyPI publish
-    needs:
-    - build
-    runs-on: ubuntu-latest
-    steps:
-    - name: Download all the dists
-      uses: actions/download-artifact@v4
-      with:
-        name: python-package-distributions-test
-        path: dist/
-    - name: List distribution files
-      run: ls -la dist/
