+++ docs/api/use_cases.md
@@ -1,9 +1,9 @@
 # Use Cases
 
-::: healthchain.use_cases.cds
+::: healthchain.sandbox.use_cases.cds
 ::: healthchain.models.requests.cdsrequest
 ::: healthchain.models.responses.cdsresponse
 
-::: healthchain.use_cases.clindoc
+::: healthchain.sandbox.use_cases.clindoc
 ::: healthchain.models.requests.cdarequest
 ::: healthchain.models.responses.cdaresponse
