+++ docs/cookbook/notereader_sandbox.md
@@ -10,7 +10,7 @@ import healthchain as hc
 from healthchain.io import Document
 from healthchain.models.requests.cda import CdaRequest, CdaResponse
 from healthchain.pipeline.medicalcodingpipeline import MedicalCodingPipeline
-from healthchain.use_cases.clindoc import ClinicalDocumentation
+from healthchain.sandbox.use_cases import ClinicalDocumentation
 from healthchain.fhir import create_document_reference
 
 from spacy.tokens import Span
