--- healthchain/clients/__init__.py
@@ -1,3 +1,0 @@
-from .ehrclient import ehr
-
-__all__ = ["ehr"]
