+++ healthchain/sandbox/clients/__init__.py
@@ -1,0 +1,3 @@
+from .ehr import EHRClient
+
+__all__ = ["EHRClient"]
