--- healthchain/use_cases/__init__.py
@@ -1,7 +1,0 @@
-from .cds import ClinicalDecisionSupport
-from .clindoc import ClinicalDocumentation
-
-__all__ = [
-    "ClinicalDecisionSupport",
-    "ClinicalDocumentation",
-]
