--- tests/interop/__init__.py
@@ -1,5 +1,0 @@
-"""
-Interop module tests
-
-Tests for the healthchain.interop module components.
-"""
